Ajout deplacerCours + Exception

+++ app/back/src/routes/cours_route.py
@@ -14,7 +14,40 @@ from psycopg2 import OperationalError, Error
 from flask_jwt_extended import JWTManager, jwt_required, create_access_token, get_jwt_identity  
 cours = Blueprint('cours', __name__)
 
+@cours.route('/cours/deplacer/<idCours>', methods=['PUT'])
+def deplacer_cours(idCours):
+    """Permet de supprimer un cours via la route /cours/delete
+    
+    :param idCours: id du cours à supprimer
+    :type idCours: int
 
+    :raises ParamètreBodyManquantException: Si aucun paramètre d'entrée attendu n'est spécifié dans le body
+    :raises ParamètreTypeInvalideException: Le type de idCours est invalide, une valeur numérique est attendue
+
+    :return: id du cours supprimer si présent
+    :rtype: json
+    """
+    if (not idCours.isdigit()):
+        return jsonify({'error': str(apiException.ParamètreTypeInvalideException("idCours", "numérique"))}), 400
+    
+    json_datas = request.get_json()
+    
+    if 'HeureDebut' not in json_datas and 'Jour' not in json_datas:
+        return jsonify({'error': str(apiException.ParamètreBodyManquantException())}), 400
+    else:
+        query = "UPDATE edt.cours "
+        if 'HeureDebut' in json_datas:
+            query += f"SET HeureDebut = '{json_datas['HeureDebut']}'"
+        if 'Jour' in json_datas:
+            if 'HeureDebut' in json_datas:
+                query += f" , Jour = '{json_datas['Jour']}'"
+            else:
+                query += f"SET Jour = '{json_datas['Jour']}'"
+        query += f" where idCours={idCours}"
+        conn = connect_pg.connect()
+        connect_pg.execute_commands(conn, query)
+        connect_pg.disconnect(conn)
+        return jsonify(idCours)
 
 @cours.route('/cours/delete/<idCours>', methods=['DELETE'])
 def delete_cours(idCours):
@@ -23,14 +56,20 @@ def delete_cours(idCours):
     :param idCours: id du cours à supprimer
     :type idCours: int
 
+    :raises ParamètreTypeInvalideException: Le type de idCours est invalide, une valeur numérique est attendue
+
     :return: id du cours supprimer si présent
     :rtype: json
     """
-    query = "delete  from edt.cours where idCours=%(idCours)s" % {'idCours':idCours}
-    conn = connect_pg.connect()
-    connect_pg.execute_commands(conn, query)
-    connect_pg.disconnect(conn)
-    return jsonify(idCours)
+    if (not idCours.isdigit()):
+        return jsonify({'error': str(apiException.ParamètreTypeInvalideException("idCours", "numérique"))}), 400
+    else:
+        query = "delete  from edt.cours where idCours=%(idCours)s" % {'idCours':idCours}
+        conn = connect_pg.connect()
+        connect_pg.execute_commands(conn, query)
+        connect_pg.disconnect(conn)
+        return jsonify(idCours)
+    
 
 @cours.route('/cours/add', methods=['POST'])
 def add_cours():
@@ -41,11 +80,14 @@ def add_cours():
     
     :raises InsertionImpossibleException: Impossible d'ajouter le cours spécifié dans la table cours
     :raises DonneeIntrouvableException: La valeur de la clée étrangère idRessource n'a pas pu être trouvé
+    :raises ParamètreBodyManquantException: Si ou plusieurs paramètres sont manquant dans le body
     
     :return: le cours qui vient d'être créé
     :rtype: json
     """
     json_datas = request.get_json()
+    if 'HeureDebut' not in json_datas or 'NombreHeure' not in json_datas or 'Jour' not in json_datas or 'idRessource' not in json_datas:
+        return jsonify({'error': str(apiException.ParamètreBodyManquantException())}), 400
     returnStatement = {}
     query = f"Insert into edt.cours (HeureDebut, NombreHeure, Jour, idRessource) values ('{json_datas['HeureDebut']}', '{json_datas['NombreHeure']}', '{json_datas['Jour']}', '{json_datas['idRessource']}') returning idCours"
     conn = connect_pg.connect()
